Task 1. If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -15,9 +15,9 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	if (array != NULL && size > 0 & action != NULL)
 	{
-		unsigned int i;
+		int i;
 
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int)size; i++)
 		{
 			action(array[i]);
 		}
